# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/mars/task": {
            "put": {
                "description": "任务更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全审计"
                ],
                "summary": "任务更新",
                "operationId": "/mars/task_put",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateSecTaskInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "添加任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全审计"
                ],
                "summary": "添加任务",
                "operationId": "/mars/task_post",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddSecTaskInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mars/task/report": {
            "get": {
                "description": "检测结果",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全审计"
                ],
                "summary": "检测结果",
                "operationId": "/mars/task/report",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "pid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mars/task_detail": {
            "get": {
                "description": "任务详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全审计"
                ],
                "summary": "任务详情",
                "operationId": "/mars/task_detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AddSecTaskInput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/mars/task_list": {
            "get": {
                "description": "任务列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "安全审计"
                ],
                "summary": "任务列表",
                "operationId": "/mars/task_list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "info",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页个数",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "当前页数",
                        "name": "page_no",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SecTaskListOutput"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/selm/task": {
            "post": {
                "description": "webhook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Web自动化测试"
                ],
                "summary": "webhook",
                "operationId": "/selm/task/post",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名",
                        "name": "project_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "项目地址",
                        "name": "project_addr",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "web地址",
                        "name": "web_addr",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "并发数",
                        "name": "sync_num",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "检索元素超时时间",
                        "name": "search_timeout",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "流水线ID",
                        "name": "stream_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/middleware.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddSecTaskInput": {
            "type": "object",
            "required": [
                "target_list",
                "task_name"
            ],
            "properties": {
                "open_exp": {
                    "type": "integer"
                },
                "open_poc": {
                    "type": "integer"
                },
                "open_web": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "scan_type": {
                    "type": "integer"
                },
                "target_list": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
...
        "dto.SecTaskListItemOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer"
                },
                "task_name": {
                    "description": "任务名",
                    "type": "string"
                },
                "task_result": {
                    "description": "检测结果",
                    "type": "integer"
                },
                "task_status": {
                    "description": "任务状态",
                    "type": "integer"
                },
                "task_targets": {
                    "description": "任务目标",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "dto.SecTaskListOutput": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SecTaskListItemOutput"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateSecTaskInput": {
            "type": "object",
            "required": [
                "id",
                "target_list",
                "task_name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "open_exp": {
                    "type": "integer"
                },
                "open_poc": {
                    "type": "integer"
                },
                "open_web": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "scan_type": {
                    "type": "integer"
                },
                "target_list": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
        "middleware.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errmsg": {
                    "type": "string"
                },
                "errno": {
                    "type": "integer"
                },
                "stack": {
                    "type": "object"
                },
                "trace_id": {
                    "type": "object"
                }
            }
        }
    }
}